requirements.txt
backcall==0.1.0
certifi==2018.8.24
chardet==3.0.4
decorator==4.3.0
Django==2.1.1
django-extensions==2.1.2
django-rest-framework==0.1.0
django-webpack-loader==0.6.0
djangorestframework==3.8.2
idna==2.7
ipdb==0.11
ipython==6.5.0
ipython-genutils==0.2.0
jedi==0.12.1
optional-django==0.3.0
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.6.0
Pygments==2.2.0
pytz==2018.5
requests==2.19.1
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
urllib3==1.23
wcwidth==0.1.7
webpack==6.0.0
